# docs/openapi_donations.yaml
...
openapi: 3.1.0
info:
  title: GoodPlay Donations API - GOO-15
  description: |
    **GOO-15 Donation Processing Engine** - Complete virtual wallet and donation system.

    ## Features
    - Virtual wallet management with credit conversion
    - Payment processing with multiple providers
    - Donation workflow with compliance checks
    - Batch processing for high-volume operations
    - Financial analytics and reporting
    - Receipt generation and tax compliance
    - Transaction reconciliation

    ## Endpoints Structure
    - `/api/wallet/*` - Virtual wallet operations
    - `/api/donations/*` - Donation processing
    - `/api/conversion-rates/*` - Credit conversion rates
    - `/api/payments/*` - Payment processing
    - `/api/batch/*` - Batch operations (admin)
    - `/api/compliance/*` - Compliance management (admin)
    - `/api/admin/financial/*` - Financial analytics (admin)

  version: 1.0.0
paths:
...
  /api/admin/financial/dashboard:
    get:
      tags:
        - Admin - Financial Analytics
      summary: Get financial dashboard
      description: Retrieve comprehensive financial KPIs, trends, and analytics
      security:
        - bearerAuth: []
        - adminRole: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Financial dashboard retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: FINANCIAL_DASHBOARD_GENERATED
                  data:
                    $ref: '#/components/schemas/FinancialDashboard'
  /api/admin/financial/analytics/users:
    get:
      tags:
        - Admin - Financial Analytics
      summary: Get user behavior analytics
      description: Retrieve user segmentation, behavior patterns, and retention metrics
      security:
        - bearerAuth: []
        - adminRole: []
      parameters:
        - name: segment
          in: query
          schema:
            type: string
            enum: [new_donors, returning_donors, high_value, frequent_donors]
      responses:
        '200':
          description: User analytics retrieved successfully
  /api/admin/financial/forecasting:
    get:
      tags:
        - Admin - Financial Analytics
      summary: Get financial forecasting
      description: Retrieve donation volume predictions and growth forecasts
      security:
        - bearerAuth: []
        - adminRole: []
      parameters:
        - name: forecast_days
          in: query
          schema:
            type: integer
            default: 7
            maximum: 90
        - name: metric
          in: query
          schema:
            type: string
            enum: [volume, count, average]
            default: volume
      responses:
        '200':
          description: Financial forecasting retrieved successfully
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminRole:
      type: apiKey
      in: header
      name: X-Admin-Role
  schemas:
...
    FinancialDashboard:
      type: object
      properties:
        core_metrics:
          type: object
          properties:
            total_volume:
              type: number
              format: float
            total_donations:
              type: integer
            average_donation:
              type: number
              format: float
            growth_rate:
              type: number
              format: float
        trends:
          type: object
          properties:
            daily_data:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  volume:
                    type: number
                    format: float
                  count:
                    type: integer
            trend_direction:
              type: string
              enum: [increasing, decreasing, stable]
        user_analytics:
          type: object
          properties:
            new_donors:
              type: integer
            returning_donors:
              type: integer
            high_value_users:
              type: integer
            retention_rate:
              type: number
              format: float
...